feat: update @tanstack/react-router and related packages; refactor order creation and detail pages
- Updated @tanstack/react-router and related packages to version 1.131.27.
- Refactored CreateOrderPage to use the new routing structure.
- Removed the old getOrder route and implemented a new dynamic route for order details.
- Created a new CreateOrderPage component to handle order creation.
- Updated the package.json and bun.lock files to reflect the new dependencies and versions.
