+++ apps/frontend/package.json
@@ -4,21 +4,21 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "dev": "bunx --bun vite",
+    "dev": "bunx --bun vite && bunx @tanstack/router-cli generate",
     "build": "tsc -b && vite build",
     "lint": "eslint .",
     "preview": "vite preview"
   },
   "dependencies": {
     "@tanstack/react-form": "^1.19.2",
-    "@tanstack/react-router": "^1.131.5",
-    "@tanstack/router-devtools": "^1.131.5",
+    "@tanstack/react-router": "^1.131.27",
     "axios": "^1.11.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.30.1",
+    "@tanstack/router-cli": "^1.131.27",
     "@tanstack/router-plugin": "^1.131.5",
     "@types/react": "^19.1.9",
     "@types/react-dom": "^19.1.7",
