+++ apps/frontend/src/routes/orders/create.tsx
@@ -1,85 +1,6 @@
-import { useForm } from "@tanstack/react-form";
-import axios from "axios";
+import { createFileRoute } from "@tanstack/react-router";
+import CreateOrderPage from "../../pages/CreateOrderPage";
 
-export const createOrderRoute = {
-  path: "/orders/create",
-  element: <CreateOrderPage />,
-};
-
-function CreateOrderPage() {
-  const form = useForm({
-    defaultValues: {
-      userId: "",
-      totalAmount: "",
-    },
-    onSubmit: async ({ value }) => {
-      try {
-        await axios.post("/api/orders", {
-          userId: value.userId,
-          totalAmount: parseFloat(value.totalAmount),
-        });
-        alert("✅ Order created successfully!");
-      } catch (err) {
-        alert("❌ Failed to create order.");
-        console.error(err);
-      }
-    },
-  });
-
-  return (
-    <div className="max-w-lg mx-auto p-8 bg-white rounded-lg shadow-lg">
-      <h1 className="text-2xl font-semibold mb-6 text-center">Create Order</h1>
-      <form
-        onSubmit={(e) => {
-          e.preventDefault();
-          form.handleSubmit();
-        }}
-        className="space-y-6"
-      >
-        <div>
-          <label className="block mb-2 font-medium text-lg text-gray-700">User ID</label>
-          <form.Field
-            name="userId"
-            children={(field) => (
-              <input
-                className="border border-gray-300 rounded-md w-full p-3 focus:outline-none focus:ring-2 focus:ring-blue-500 transition"
-                placeholder="Enter User ID"
-                value={field.state.value}
-                onBlur={field.handleBlur}
-                onChange={(e) => field.handleChange(e.target.value)}
-              />
-            )}
-          />
-        </div>
-
-        <div>
-          <label className="block mb-2 font-medium text-lg text-gray-700">Total Amount</label>
-          <form.Field
-            name="totalAmount"
-            children={(field) => (
-              <input
-                type="number"
-                className="border border-gray-300 rounded-md w-full p-3 focus:outline-none focus:ring-2 focus:ring-blue-500 transition"
-                placeholder="Enter Total Amount"
-                value={field.state.value}
-                onBlur={field.handleBlur}
-                onChange={(e) => field.handleChange(e.target.value)}
-              />
-            )}
-          />
-        </div>
-
-        <div className="flex justify-center">
-          <button
-            type="submit"
-            className="bg-blue-600 text-white px-6 py-3 rounded-md font-medium shadow-md hover:bg-blue-700 transition duration-200"
-          >
-            Create Order
-          </button>
-        </div>
-      </form>
-    </div>
-  );
-}
-
-export default CreateOrderPage;
+export const Route = createFileRoute('/orders/create')({
+  component: CreateOrderPage
+});
